backend/requirements.txt
Flask==3.1.0
Flask-CORS==4.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Werkzeug>=3.1,<4
marshmallow==3.22.0
joblib==1.4.2
scikit-learn==1.5.2
numpy==2.1.3
pytest==8.3.3
pytest-flask==1.3.0
Flask-Limiter==3.8.0
idna==3.11

